+++ CHANGES.md
@@ -48,7 +48,8 @@ To be released.
  -  Added `IStore.ForkStateReferences<T>(string, string, Block<T>,
     IImmutableSet<Address>` method.  [[#232]]
  -  Removed `Block<T>.Validate()` and `Block<T>.EvaluateActions()` method.  [[#243]]
- -  Added `Transaction<T>.Nonce` and `RawTransaction.Nonce`.  [[#246]]
+ -  Added `Transaction<T>.Nonce` and `RawTransaction.Nonce` properties.
+    [[#246]]
  -  Added `IStore.GetTxNonce(string, Address)` method.  [[#246]]
  -  Added `IStore.IncreaseTxNonce<T>(string, Block<T>)` method.  [[#246]]
  -  Added `IStore.ForkTxNonce<T>(string, string, Block<T>,
@@ -109,8 +110,8 @@ To be released.
  -  `BlockChain<T>.Append()` became to execute `Action.Execute()` only once per
     action in the `Block<T>`.  [[#243]]
  -  `BlockChain<T>.Append()` method became to throw `InvalidTxNonceException`
-    when the `Transaction<T>.Nonce` is different from `BlockChain<T>.GetNonce`
-    result of the `Transaction<T>.Signer`.  [[#246]]
+    when the `Transaction<T>.Nonce` does not correspond to its `Signer`'s
+    current nonce.  [[#246]]
 
 ### Bug fixes